requirements.txt
google-genai>=1.0.0
feedparser>=6.0.11
requests>=2.31.0
tweepy>=4.14.0
Pillow>=10.0.0
